[ADD] purchase_sale_inter_company_stock_intercompany: Glue module
