+++ purchase_sale_inter_company/tests/common.py
@@ -1,0 +1,163 @@
+# Copyright 2013-Today Odoo SA
+# Copyright 2019-2019 Chafique DELLI @ Akretion
+# Copyright 2018-2019 Tecnativa - Carlos Dauden
+# Copyright 2020 ForgeFlow S.L. (https://www.forgeflow.com)
+# License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
+
+from odoo.tests.common import Form
+
+from odoo.addons.account_invoice_inter_company.tests.test_inter_company_invoice import (
+    TestAccountInvoiceInterCompanyBase,
+)
+
+
+class TestPurchaseSaleInterCompanyCommon(TestAccountInvoiceInterCompanyBase):
+    @classmethod
+    def _create_warehouse(cls, code, company):
+        address = cls.env["res.partner"].create({"name": f"{code} address"})
+        return cls.env["stock.warehouse"].create(
+            {
+                "name": f"Warehouse {code}",
+                "code": code,
+                "partner_id": address.id,
+                "company_id": company.id,
+            }
+        )
+
+    @classmethod
+    def _configure_user(cls, user):
+        for xml in [
+            "account.group_account_manager",
+            "base.group_partner_manager",
+            "sales_team.group_sale_manager",
+            "purchase.group_purchase_manager",
+        ]:
+            user.groups_id |= cls.env.ref(xml)
+
+    @classmethod
+    def _create_purchase_order(cls, partner, product_id=None):
+        po = Form(cls.env["purchase.order"])
+        po.company_id = cls.company_a
+        po.partner_id = partner
+
+        cls.product.invoice_policy = "order"
+
+        with po.order_line.new() as line_form:
+            line_form.product_id = product_id if product_id else cls.product
+            line_form.product_qty = 3.0
+            line_form.name = "Service Multi Company"
+            line_form.price_unit = 450.0
+        return po.save()
+
+    @classmethod
+    def setUpClass(cls):
+        super().setUpClass()
+        cls.lot_obj = cls.env["stock.production.lot"]
+        cls.quant_obj = cls.env["stock.quant"]
+        # no job: avoid issue if account_invoice_inter_company_queued is installed
+        cls.env = cls.env(context={"test_queue_job_no_delay": 1})
+
+        cls.product = cls.product_consultant_multi_company
+
+        cls.consumable_product = cls.env["product.product"].create(
+            {
+                "name": "Consumable Product",
+                "type": "consu",
+                "categ_id": cls.env.ref("product.product_category_all").id,
+                "qty_available": 100,
+            }
+        )
+        cls.consumable_product_2 = cls.env["product.product"].create(
+            {
+                "name": "Consumable Product 2",
+                "type": "consu",
+            }
+        )
+        cls.stockable_product_serial = cls.env["product.product"].create(
+            {
+                "name": "Stockable Product Tracked by Serial",
+                "type": "product",
+                "tracking": "serial",
+                "categ_id": cls.env.ref("product.product_category_all").id,
+            }
+        )
+
+        # if partner_multi_company or product_multi_company is installed
+        # We have to do that because the default method added a company
+        if "company_ids" in cls.env["res.partner"]._fields:
+            cls.partner_company_a.company_ids = False
+            cls.partner_company_b.company_ids = False
+
+        if "company_ids" in cls.env["product.template"]._fields:
+            cls.product.company_ids = False
+            cls.consumable_product.company_ids = False
+
+        # Configure 2 Warehouse per company
+        cls.warehouse_a = cls.env["stock.warehouse"].search(
+            [("company_id", "=", cls.company_a.id)]
+        )
+        cls.warehouse_b = cls._create_warehouse("CA-WB", cls.company_a)
+
+        cls.warehouse_c = cls.env["stock.warehouse"].search(
+            [("company_id", "=", cls.company_b.id)]
+        )
+        cls.warehouse_d = cls._create_warehouse("CB-WD", cls.company_b)
+
+        # Configure Company B (the supplier)
+        cls.company_b.so_from_po = True
+        cls.company_b.warehouse_id = cls.warehouse_c
+        cls.company_b.sale_auto_validation = 1
+
+        cls.intercompany_sale_user_id = cls.user_company_b.copy()
+        cls.intercompany_sale_user_id.company_ids |= cls.company_a
+        cls.company_b.intercompany_sale_user_id = cls.intercompany_sale_user_id
+
+        # Configure User
+        cls._configure_user(cls.user_company_a)
+        cls._configure_user(cls.user_company_b)
+
+        # Create purchase order
+        cls.purchase_company_a = cls._create_purchase_order(cls.partner_company_b)
+
+        # Configure pricelist to USD
+        cls.env["product.pricelist"].sudo().search([]).write(
+            {"currency_id": cls.env.ref("base.USD").id}
+        )
+
+        # Add quants for product tracked by serial to supplier
+        cls.serial_1 = cls._create_serial_and_quant(
+            cls.stockable_product_serial, "111", cls.company_b
+        )
+        cls.serial_2 = cls._create_serial_and_quant(
+            cls.stockable_product_serial, "222", cls.company_b
+        )
+        cls.serial_3 = cls._create_serial_and_quant(
+            cls.stockable_product_serial, "333", cls.company_b
+        )
+
+    @classmethod
+    def _create_serial_and_quant(cls, product, name, company, quant=True):
+        lot = cls.lot_obj.create(
+            {"product_id": product.id, "name": name, "company_id": company.id}
+        )
+        if quant:
+            cls.quant_obj.create(
+                {
+                    "product_id": product.id,
+                    "location_id": cls.warehouse_a.lot_stock_id.id,
+                    "quantity": 1,
+                    "lot_id": lot.id,
+                }
+            )
+        return lot
+
+    def _approve_po(self, purchase_id):
+        """Confirm the PO in company A and return the related sale of Company B"""
+
+        purchase_id.with_user(self.intercompany_sale_user_id).button_approve()
+
+        return (
+            self.env["sale.order"]
+            .with_user(self.user_company_b)
+            .search([("auto_purchase_order_id", "=", purchase_id.id)])
+        )
