+++ purchase_sale_inter_company_stock_intercompany/__manifest__.py
@@ -1,0 +1,17 @@
+# Copyright 2022 Akretion
+# License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
+
+{
+    "name": "Compatibility module for Purchase Sale Inter Company and Stock Intercompany",
+    "version": "14.0.1.0.0",
+    "category": "Purchase Management",
+    "website": "https://github.com/OCA/multi-company",
+    "author": "Akretion, Odoo Community Association (OCA)",
+    "license": "AGPL-3",
+    "installable": True,
+    "depends": [
+        "purchase_sale_inter_company",
+        "stock_intercompany",
+    ],
+    "auto_install": True,
+}
