+++ purchase_sale_inter_company_stock_intercompany/models/__init__.py
@@ -1,0 +1,1 @@
+from . import stock_picking
