+++ purchase_sale_inter_company_stock_intercompany/readme/CONFIGURATION.rst
@@ -1,0 +1,1 @@
+Configuration is automatic
