+++ purchase_sale_inter_company_stock_intercompany/readme/CONTRIBUTORS.rst
@@ -1,0 +1,1 @@
+* Florian Mounier <florian.mounier@akretion.com>
