+++ purchase_sale_inter_company_stock_intercompany/readme/DESCRIPTION.rst
@@ -1,0 +1,1 @@
+Compatibility module for Purchase Sale Inter Company and Stock Intercompany
