+++ purchase_sale_inter_company_stock_intercompany/tests/__init__.py
@@ -1,0 +1,1 @@
+from . import test_inter_company_purchase_sale_stock_intercompany
