+++ setup/purchase_sale_inter_company_stock_intercompany/odoo/addons/purchase_sale_inter_company_stock_intercompany
@@ -1,0 +1,1 @@
+../../../../purchase_sale_inter_company_stock_intercompany
