+++ stock_intercompany/tests/test_intercompany_picking.py
@@ -1,91 +1,13 @@
 # Copyright 2021 Camptocamp
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl.html).
 
-from odoo.tests.common import SavepointCase
 from odoo.tests import tagged
+
+from .common import TestStockIntercompanyCommon
 
 
 @tagged("post_install", "-at_install")
-class TestStockIntercompany(SavepointCase):
-    @classmethod
-    def setUpClass(cls):
-        super().setUpClass()
-        cls.user_demo = cls.env.ref("base.user_demo")
-        company_obj = cls.env["res.company"]
-        # Create 2 companies and configure intercompany picking type param on them
-        cls.company1 = company_obj.search(
-            [("name", "=", "Company A")]
-        ) or company_obj.create({"name": "Company A"})
-        cls.company2 = company_obj.search(
-            [("name", "=", "Company B")]
-        ) or company_obj.create({"name": "Company B"})
-        cls.picking_type_out_company1 = (
-            cls.env["stock.picking.type"]
-            .sudo()
-            .search(
-                [
-                    ("company_id", "=", cls.company1.id),
-                    ("name", "=", "Delivery Orders"),
-                ],
-                limit=1,
-            )
-        )
-        cls.picking_type_out_company2 = (
-            cls.env["stock.picking.type"]
-            .sudo()
-            .search(
-                [
-                    ("company_id", "=", cls.company2.id),
-                    ("name", "=", "Delivery Orders"),
-                ],
-                limit=1,
-            )
-        )
-        cls.picking_type_in_company1 = (
-            cls.env["stock.picking.type"]
-            .sudo()
-            .search(
-                [("company_id", "=", cls.company1.id), ("name", "=", "Receipts")],
-                limit=1,
-            )
-        )
-        cls.picking_type_in_company2 = (
-            cls.env["stock.picking.type"]
-            .sudo()
-            .search(
-                [("company_id", "=", cls.company2.id), ("name", "=", "Receipts")],
-                limit=1,
-            )
-        )
-
-        cls.company1.intercompany_in_type_id = cls.picking_type_in_company1.id
-        cls.company2.intercompany_in_type_id = cls.picking_type_in_company2.id
-
-        cls.company1.intercompany_out_type_id = cls.picking_type_out_company1.id
-        cls.company2.intercompany_out_type_id = cls.picking_type_out_company2.id
-        # assign both companies to current user
-        cls.user_demo.write(
-            {
-                "company_id": cls.company1.id,
-                "company_ids": [(4, cls.company1.id), (4, cls.company2.id)],
-            }
-        )
-        # create storable product
-        cls.product1 = cls.env["product.product"].create(
-            {
-                "name": "Product A",
-                "type": "product",
-                "categ_id": cls.env.ref("product.product_category_all").id,
-                "qty_available": 100,
-            }
-        )
-        cls.stock_location = (
-            cls.env["stock.location"]
-            .sudo()
-            .search([("name", "=", "Stock"), ("company_id", "=", cls.company1.id)])
-        )
-        cls.uom_unit = cls.env.ref("uom.product_uom_unit")
-
+class TestStockIntercompany(TestStockIntercompanyCommon):
     def test_picking_creation_same_company(self):
         picking = (
             self.env["stock.picking"]
@@ -526,3 +448,56 @@ class TestStockIntercompany(SavepointCase):
             ),
             0,
         )
+
+    def test_picking_out_cancel(self):
+        self.company2.intercompany_picking_creation_mode = "out"
+        # Create an out when an in is created
+
+        self.assertEquals(
+            self.env["stock.picking"]
+            .with_context(default_company_id=self.company2.id)
+            .with_user(self.user_demo)
+            .sudo()
+            .search_count(
+                [("picking_type_id", "=", self.picking_type_out_company2.id)]
+            ),
+            0,
+        )
+
+        picking = (
+            self.env["stock.picking"]
+            .with_context(default_company_id=self.company1.id)
+            .with_user(self.user_demo)
+            .create(
+                {
+                    "partner_id": self.company2.partner_id.id,
+                    "location_id": self.env.ref("stock.stock_location_suppliers").id,
+                    "location_dest_id": self.stock_location.id,
+                    "picking_type_id": self.picking_type_in_company1.id,
+                }
+            )
+        )
+
+        self.env["stock.move.line"].create(
+            {
+                "location_id": self.stock_location.id,
+                "location_dest_id": self.env.ref("stock.stock_location_suppliers").id,
+                "product_id": self.product1.id,
+                "product_uom_id": self.uom_unit.id,
+                "qty_done": 1.0,
+                "picking_id": picking.id,
+            }
+        )
+
+        picking.action_confirm()
+
+        counterpart_picking = (
+            self.env["stock.picking"]
+            .with_context(default_company_id=self.company2.id)
+            .with_user(self.user_demo)
+            .sudo()
+            .search([("picking_type_id", "=", self.picking_type_out_company2.id)])
+        )
+        picking.action_cancel()
+        self.assertEqual(picking.state, "cancel")
+        self.assertEqual(counterpart_picking.state, "cancel")
